style: remove isort ruff style
