+++ pyproject.toml
@@ -81,8 +81,9 @@ select = [
     "W",
 ]
 
-[tool.ruff.lint.isort]
-lines-after-imports = 2
+
+# [tool.ruff.lint.isort]
+# lines-after-imports = 2
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = [
